Revert "merge into console.scala"
This reverts commit 8ce6f3881a1dabbc6cebf2702960961571fce00e.
