+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/sources/ConsoleWriter.scala
@@ -17,19 +17,14 @@
 
 package org.apache.spark.sql.execution.streaming.sources
 
-import java.util.Optional
-
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql._
-import org.apache.spark.sql.sources.{BaseRelation, CreatableRelationProvider, DataSourceRegister}
-import org.apache.spark.sql.sources.v2.{DataSourceV2, DataSourceV2Options}
-import org.apache.spark.sql.sources.v2.streaming.MicroBatchWriteSupport
+import org.apache.spark.sql.{Row, SparkSession}
+import org.apache.spark.sql.sources.v2.DataSourceV2Options
 import org.apache.spark.sql.sources.v2.writer.{DataSourceV2Writer, DataWriterFactory, WriterCommitMessage}
-import org.apache.spark.sql.streaming.OutputMode
 import org.apache.spark.sql.types.StructType
 
 class ConsoleWriter(batchId: Long, schema: StructType, options: DataSourceV2Options)
-  extends DataSourceV2Writer with Logging {
+    extends DataSourceV2Writer with Logging {
   // Number of rows to display, by default 20 rows
   private val numRowsToShow = options.getInt("numRows", 20)
 
@@ -59,41 +54,4 @@ class ConsoleWriter(batchId: Long, schema: StructType, options: DataSourceV2Opti
   override def abort(messages: Array[WriterCommitMessage]): Unit = {}
 
   override def toString(): String = s"ConsoleWriter[numRows=$numRowsToShow, truncate=$isTruncated]"
-}
-
-case class ConsoleRelation(override val sqlContext: SQLContext, data: DataFrame)
-  extends BaseRelation {
-  override def schema: StructType = data.schema
-}
-
-class ConsoleSinkProvider extends DataSourceV2
-  with MicroBatchWriteSupport
-  with DataSourceRegister
-  with CreatableRelationProvider {
-
-  override def createMicroBatchWriter(
-      queryId: String,
-      epochId: Long,
-      schema: StructType,
-      mode: OutputMode,
-      options: DataSourceV2Options): Optional[DataSourceV2Writer] = {
-    Optional.of(new ConsoleWriter(epochId, schema, options))
-  }
-
-  def createRelation(
-      sqlContext: SQLContext,
-      mode: SaveMode,
-      parameters: Map[String, String],
-      data: DataFrame): BaseRelation = {
-    // Number of rows to display, by default 20 rows
-    val numRowsToShow = parameters.get("numRows").map(_.toInt).getOrElse(20)
-
-    // Truncate the displayed data if it is too long, by default it is true
-    val isTruncated = parameters.get("truncate").map(_.toBoolean).getOrElse(true)
-    data.show(numRowsToShow, isTruncated)
-
-    ConsoleRelation(sqlContext, data)
-  }
-
-  def shortName(): String = "console"
 }
